% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_gene_sets.R
\name{process_gene_sets}
\alias{process_gene_sets}
\title{Process and Filter Gene Sets}
\usage{
process_gene_sets(
  filename,
  seurat_object,
  output_folder_base = output_folder_base,
  min_genes = 10,
  max_genes = 100
)
}
\arguments{
\item{filename}{A string specifying the path to the gene sets file.}

\item{seurat_object}{A Seurat object containing the RNA assays.}

\item{output_folder_base}{A string specifying the base path where the output file will be saved. Default: output_folder_base.}

\item{min_genes}{An integer giving the minimum number of genes for a gene set to be included. Default: 10.}

\item{max_genes}{An integer giving the maximum number of genes for a gene set to be included. Default: 100.}
}
\value{
A data frame of gene sets where each column represents a gene set, and each row
represents a gene. The data frame contains binary values indicating gene presence in each set.
}
\description{
Reads an enrichR gene - gene set library file, processes each line to extract gene sets and their genes,
and filters these gene sets based on the gene count criteria. It integrates with a Seurat object
to match and filter the gene sets. Output can be used for a scSpectra analysis.
}
\details{
https://maayanlab.cloud/Enrichr/#libraries
}
\examples{
# Assuming you have a valid 'seurat_object' and a gene sets file:
#go_sets <- process_gene_sets(
#filename = "path/to/file.txt",
#seurat_object = object,
#output_folder_base = output_folder_base,
#min_genes = 10,
#max_genes = 100)

}
