% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_gene_sets_for_matrix.R
\name{process_gene_sets_for_matrix}
\alias{process_gene_sets_for_matrix}
\title{Process Gene Sets (for a Matrix)}
\usage{
process_gene_sets_for_matrix(filename, datD, min_genes = 10, max_genes = 100)
}
\arguments{
\item{filename}{A string, the path to the file containing gene sets.}

\item{datD}{A matrix, typically a pseudobulk matrix, where cols are genes.}

\item{min_genes}{An integer, the minimum number of genes for a gene set to be considered valid.}

\item{max_genes}{An integer, the maximum number of genes for a gene set to be considered valid.}
}
\value{
A data frame where each column represents a gene set, rows represent genes,
        and entries are binary (1 if the gene is in the gene set, 0 otherwise).
        The data frame is filtered based on the specified gene count criteria.
}
\description{
Reads a file containing gene sets, processes each line to extract gene sets and their genes,
then matches these with genes in the provided matrix. The function filters the gene sets
based on a minimum and maximum number of genes. The difference to the process_gene_sets function
is the matching to a (pseudo)bulk matrix of samples (rows) and genes (columns).
}
\details{
https://maayanlab.cloud/Enrichr/#libraries

or

https://www.gsea-msigdb.org/gsea/msigdb
}
\examples{
#gene_sets_file <- "path/to/gene_sets.txt"
#pseudobulk_matrix <- matrix(data, nrow = number_of_samples, ncol = number_of_genes)
#go_sets <- process_gene_sets_for_matrix(gene_sets_file, datD = pseudobulk_matrix, 10, 100)

}
