% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_gene_sets_from_GO.R
\name{process_gene_sets_from_GO}
\alias{process_gene_sets_from_GO}
\title{Process Gene Sets from GO Annotations}
\usage{
process_gene_sets_from_GO(
  gaf_file_path,
  obo_file_path,
  genes,
  min_genes = 10,
  max_genes = 100
)
}
\arguments{
\item{gaf_file_path}{String; the file path to the GAF file containing GO annotations. Make sure to download file for species of interest.}

\item{obo_file_path}{String; the file path to the OBO file containing GO term names.}

\item{genes}{Vector of strings; the list of gene symbols to include in the analysis. e.g. gene symbols in a single cell or gene expression matrix.}

\item{min_genes}{Integer; the minimum number of genes for a GO set to be included (default = 10).}

\item{max_genes}{Integer; the maximum number of genes for a GO set to be included (default = 100).}
}
\value{
A binary matrix where rows represent genes, columns represent GO sets (with GO IDs
        and names), and values are 1 (gene is in the GO set) or 0 (gene is not in the GO set).
        The matrix includes only GO sets with gene counts within the specified range.
}
\description{
Reads Gene Ontology (GO) annotations from a GAF file and GO term names from an OBO file. It then
creates a binary matrix indicating the presence (1) or absence (0) of genes in each GO set.
The matrix is filtered based on a specified minimum and maximum number of genes.
}
\details{
latest files can be downloaded from: http://current.geneontology.org
}
\examples{
# Assuming you have the necessary files and a list of genes
# gaf_file <- "path/to/gaf_file.gaf"
# obo_file <- "path/to/obo_file.obo"
# genes <- c("gene1", "gene2", "gene3")
# go_sets <- process_gene_sets_from_GO(gaf_file, obo_file, genes, 10, 100)

}
