% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_genes.R
\name{report_genes}
\alias{report_genes}
\title{Report Genes}
\usage{
report_genes(
  output_folder_base,
  gene_set_name,
  cell_type,
  sample_of_interest,
  expression_threshold = 0
)
}
\arguments{
\item{output_folder_base}{Directory for loading the reference and query data files.}

\item{gene_set_name}{Name of the gene set or pathway of interest.}

\item{cell_type}{Cell type to analyze.}

\item{sample_of_interest}{Specific sample to analyze.}

\item{expression_threshold}{Threshold to filter genes of low expression.}
}
\value{
A dataframe with genes and their normalized residuals.

#' @examples
#gene_set_name = "Postsynaptic Membrane Organization (GO:0001941)"
#output_folder_base = "/.../output_human_extended/"
#drivinggenes <- report_genes(output_folder_base, gene_set_name, cell_type = "TAL",
#sample_of_interest = "sample")
}
\description{
This function identifies driver genes for a specific cell type, function, and sample of interest.
It calculates residuals for each gene, normalized by the reference LOO distribution, and returns a
dataframe of genes and respective "normalized residuals" above the expression threshold
}
