% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_genes_from_matrix.R
\name{report_genes_from_matrix}
\alias{report_genes_from_matrix}
\title{Report Genes from Matrix}
\usage{
report_genes_from_matrix(
  datH,
  datD,
  go_sets,
  gene_set_name,
  sample_of_interest,
  expression_threshold = 0
)
}
\arguments{
\item{datH}{Reference matrix (rows = samples, columns = genes) for healthy/control data.}

\item{datD}{Query matrix (rows = samples, columns = genes), which should include the
sample_of_interest for driver gene identification.}

\item{go_sets}{Matrix or data frame indicating membership of genes in gene sets. Rows should be
gene names, and columns should be specific gene sets. A value of 1 indicates membership
in that gene set.}

\item{gene_set_name}{Name of the gene set (function/pathway) of interest. This should match one
of the column names of the gene set database.}

\item{sample_of_interest}{Specific sample (row in datD) to analyze.}

\item{expression_threshold}{Threshold to filter genes of low expression. Genes must exceed this
average expression level in both datH and datD.}
}
\value{
A dataframe with genes and their normalized residuals.

#' @examples
# gene_set_name = "Postsynaptic Membrane Organization (GO:0001941)"
# sample_of_interest = "sample_1"
#
# driver_genes_from_matrix <- report_genes_from_matrix(
#   datH = reference_matrix,
#   datD = query_matrix,
#   go_sets = go_sets,
#   gene_set_name = gene_set_name,
#   sample_of_interest = sample_of_interest,
#   expression_threshold = 0
# )
}
\description{
This function identifies underlying genes for a specific function (gene set) and sample of interest
from a preprocessed matrix, instead of from a Seurat object. It calculates residuals for each
gene, normalized by the reference leave-one-out (LOO) distribution, and returns a dataframe of
genes and their respective "normalized residuals" above the expression threshold.
}
