% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_spectra_from_matrix.R
\name{run_spectra_from_matrix}
\alias{run_spectra_from_matrix}
\title{Run Spectra Analysis from Matrix Data}
\usage{
run_spectra_from_matrix(
  datH,
  datD,
  go_sets,
  output_folder,
  coordination_key = "nokey",
  CHISQ.MAX = 4
)
}
\arguments{
\item{datH}{A normalized (e.g. tpm + log) matrix of gene expression data for healthy samples, with genes as rows.}

\item{datD}{A normalized (e.g. tpm + log) matrix of gene expression data for diseased samples, similar to datH.}

\item{go_sets}{A dataframe or matrix where columns represent gene sets. Can be generated using process_gene_sets_for_matrix function.}

\item{output_folder}{The directory path where the result files will be saved.}

\item{coordination_key}{An optional key for file naming (default is 'nokey'). Could be cell type or condition.}

\item{CHISQ.MAX}{An optional maximum chi-square value for robustness (default is 4).}
}
\value{
This function does not return a value but saves two CSV files containing R² and p-values.
}
\description{
Performs spectra analysis on given matrices query (datD) samples compared to a reference (datH).
It calculates R² and p-values for each gene set and saves the results as CSV files.
}
\examples{
# datH and datD are matrices with the same genes but different samples
#(healthy/reference vs diseased/query)
#run_spectra_from_matrix(datH, datD, go_sets, "path/to/output", "key_example")

}
