% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_references.R
\name{create_references}
\alias{create_references}
\title{Create Pseudobulk References Files from Seurat Object matching the counts of query samples.}
\usage{
create_references(
  seurat_object,
  output_folder_base = output_folder_base,
  num_replicates = 1,
  cell_types = cell_types,
  cell_number_threshold = 10,
  seed = NULL
)
}
\arguments{
\item{seurat_object}{A Seurat object to process.}

\item{output_folder_base}{A string specifying the base path where output files will be saved. Default: output_folder_base.}

\item{num_replicates}{The number of replicates to generate for each reference sample.}

\item{cell_types}{A vector of cell types to process. Default: cell_types.}

\item{cell_number_threshold}{Minimum number of cells required to consider a sample as valid. Default = 10.}

\item{seed}{Optional; an integer seed for reproducibility of random processes.}
}
\value{
The function does not return a value but saves output files (datH and datD matrices) in specified directories.
}
\description{
This is the "on the fly" input for our spectra analysis.
It ensures that differences in cell numbers between query and reference samples drive an inflation of dyscoordination.
}
\details{
This function processes a Seurat object to create pseudobulk references. It subsets the Seurat object
by cell type, identifies valid samples based on a cell number threshold, and generates pseudobulk
references for both healthy and diseased samples. Results are saved as RDS files in specified output folders.
}
\examples{
# Assuming 'seurat_object' is your Seurat object and 'output_folder_base' is defined:
#create_references(seurat_object,
#output_folder_base,
#num_replicates = 1,
#cell_types = c("CellType1", "CellType2"),
#cell_number_threshold = 10,
#seed = NULL)

}
