% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_seurat_object.R
\name{prepare_seurat_object}
\alias{prepare_seurat_object}
\title{Prepare Seurat Object for reference generation / pseudobulking / spectra analysis}
\usage{
prepare_seurat_object(
  seurat_obj,
  celltype_col,
  sample_id_col,
  condition_col,
  query_list,
  control_list
)
}
\arguments{
\item{seurat_obj}{A Seurat object to be modified.}

\item{celltype_col}{Name of the column in the Seurat object metadata that contains cell type annotations.}

\item{sample_id_col}{Name of the column in the Seurat object metadata that contains sample/patient identifiers.}

\item{condition_col}{Name of the column in the Seurat object metadata that contains condition information.}

\item{query_list}{List of values in the condition column to be mapped as 'query'.}

\item{control_list}{List of values in the condition column to be mapped as 'control'.}
}
\value{
A modified Seurat object with additional metadata columns: celltype_annotation, orig_ident, and query_control.
}
\description{
Modifies a Seurat object to include specific metadata: cell type annotation, sample/patient identifier,
and condition information (mapped to 'query' or 'control' based on provided lists).
}
\examples{
# Assuming you have a Seurat object 'seurat_obj':
#modified_seurat <- prepare_seurat_object(seurat_obj,
#"cellType",
#"patientID",
#"condition",
#c("diseased1", "diseased2"),
#c("Control"))

}
