% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_gene_sets_from_KEGG.R
\name{process_gene_sets_from_KEGG}
\alias{process_gene_sets_from_KEGG}
\title{Process_gene_sets_from_KEGG}
\usage{
process_gene_sets_from_KEGG(organism, seurat_object)
}
\arguments{
\item{organism}{Character. The KEGG organism code (e.g., "mmu" for mouse).}

\item{seurat_object}{Seurat object used for the spectra analysis.}
}
\value{
A dataframe where rows are genes and columns are KEGG pathways.
        Each cell contains 1 if the gene is part of the pathway, otherwise 0.
}
\description{
This function retrieves all KEGG pathways for a given organism and creates a binary matrix
indicating the presence of genes (represented in a Seurat object) in these pathways. The
matrix has to be saved manually to the output base folder ("go_sets_modified.rds")
}
\examples{

organism <- "mmu"

#kegg_gene_matrix <- process_gene_sets_from_KEGG(organism, seurat_object = seurat_object)
#saveRDS(kegg_gene_matrix, "go_sets_modified.rds")
}
