% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_spectra.R
\name{run_spectra}
\alias{run_spectra}
\title{Run Spectra Analysis}
\usage{
run_spectra(
  output_folder_base = output_folder_base,
  cell_types = cell_types,
  CHISQ.MAX = 4,
  expression_threshold = 0,
  gene_number_threshold = 10,
  restrict_to_sample = FALSE,
  results_sample_key = "sampleofinterest",
  vector_of_samples = c(),
  QC_report = TRUE
)
}
\arguments{
\item{output_folder_base}{The base directory for saving output files. The default is set to the
value of the 'output_folder_base' variable defined outside the function.}

\item{cell_types}{Vector of cell types to be processed, defaulting to the 'cell_types' variable
defined outside the function.}

\item{CHISQ.MAX}{Maximum value for chi-square statistic used to cut-off the contribution of an individual gene of a pathway, with a default
value of 4.}

\item{expression_threshold}{Genes with an average expression in the reference
OR query matrix below this threshold are removed from the analysis.
Default 0.}

\item{gene_number_threshold}{Number of genes in a gene set after filtering needed to perform analysis. If less genes remaining, results will be NA.
Default 10 genes.}

\item{restrict_to_sample}{Set TRUE if you want to specify the samples to run
scSpectra on. If FALSE, the tool will run analysis based on the
valid_samples file created by the create_references function.
Make sure to specify the vector_of_samples.
Default: FALSE.}

\item{results_sample_key}{Needed if restrict_to_sample = TRUE. Will specify
the samplekey to use when generating R2 and Pval results into the
output folder. Default: "sampleofinterest".}

\item{vector_of_samples}{A vector of sample in the shape of
c("sample1", "sample2"). Only needed if restrict_to_sample = TRUE.}

\item{QC_report}{If TRUE a QC report of the reference coordination for each
sample / cell type and every pathway will be generated (including the
mean, median, and sd of the reference LOO R2 distribution and the
variance explained by V1 of the reference).
Default: TRUE.}
}
\value{
Does not return any value, but outputs files to the specified directory.
}
\description{
This function performs the actual scSpectra analysis on our created references from a seurat object, calculates R² and p-values for each gene set, and saves the results in
specified subfolders ('R2' and 'Pval') within the output directory. Only needs the output folder from the create_references function and the cell types to analyze.
}
\examples{
# Assuming 'seurat_object' is defined and 'output_folder_base' and 'cell_types' are set:
#run_spectra(output_folder_base = output_folder_base,
#cell_types = cell_types, CHISQ.MAX = 4, expression_threshold = 0,
#gene_number_threshold = 10, restrict_to_sample = FALSE,
#vector_of_samples = c())


}
